+++ src/main/java/org/zkoss/essentials/chapter2/SidebarPageConfigChapter2Impl.java
@@ -6,7 +6,7 @@
 
 Copyright (C) 2012 Potix Corporation. All Rights Reserved.
 */
-package org.zkoss.essentials.chapter4;
+package org.zkoss.essentials.chapter2;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -16,10 +16,10 @@ import java.util.List;
 import org.zkoss.essentials.services.SidebarPage;
 import org.zkoss.essentials.services.SidebarPageConfig;
 
-public class SidebarPageConfigChapter4Impl implements SidebarPageConfig{
+public class SidebarPageConfigChapter2Impl implements SidebarPageConfig{
 
 	HashMap<String,SidebarPage> pageMap = new LinkedHashMap<String,SidebarPage>();
-	public SidebarPageConfigChapter4Impl(){		
+	public SidebarPageConfigChapter2Impl(){		
 		pageMap.put("fn1",new SidebarPage("zk","ZK","/imgs/site.png","http://www.zkoss.org/"));
 		pageMap.put("fn2",new SidebarPage("demo","ZK Demo","/imgs/demo.png","http://www.zkoss.org/zkdemo"));
 		pageMap.put("fn3",new SidebarPage("devref","ZK Developer Reference","/imgs/doc.png","http://books.zkoss.org/wiki/ZK_Developer's_Reference"));
