+++ src/main/webapp/chapter2/index.zul
@@ -5,7 +5,7 @@
 			<include src="/chapter1/banner.zul"/>
 		</north>
 		<west width="260px" border="none" collapsible="true" splittable="true" minsize="300">
-			<include src="/chapter4/sidebar.zul"/>
+			<include src="/chapter2/sidebar.zul"/>
 		</west>
 		<center id="mainContent" autoscroll="true">
 			<vbox vflex="1" hflex="1" align="center" pack="center" spacing="20px" >
