+++ src/main/webapp/chapter2/sidebar.zul
@@ -1,6 +1,6 @@
 <grid hflex="1" vflex="1" sclass="sidebar" 
 	id="fnList" 
-	apply="org.zkoss.essentials.chapter4.SidebarChapter4Controller">
+	apply="org.zkoss.essentials.chapter2.SidebarChapter2Controller">
 	<columns>
 		<column width="36px"/>
 		<column/>
