+++ src/main/webapp/chapter6/index-mvvm.zul
@@ -5,7 +5,7 @@
 			<include src="/chapter1/banner.zul"/>
 		</north>
 		<west width="260px" border="none" collapsible="true" splittable="true" minsize="300">
-			<include src="/chapter4/sidebar.zul"/>
+			<include src="/chapter2/sidebar.zul"/>
 		</west>
 		<center id="mainContent" autoscroll="true" border="none">
 			<include src="/chapter6/todolist-mvvm.zul"/>
